wm/xdg_client: always refocus when resize event finishes
We set up a timeout of 50ms when clients get resized during which we
stop drawing frames. This is to avoid drawing imperfect frames to the
best extent possible.

But we use this timeout for a second purpose, as well: if the timeout
hits we will immediately refocus the client regardless of whether it has
resized already or not in order to update the focus. This is wrong
though: we should only ever update the focus once the resize has been
handled.

Start treating these as two different things:

  - We use the `pending_serial` to track whether the resize has
    finished, which will causue us to both stop inhibiting frames and to
    re-focus the client.

  - We use the timeout to stop inhibiting frames, only.

Like this, we do the correct thing regardless of whether or not the
timeout kicks in.