wm/output: reapply gamma when powering on output
Starting with wlroots 0.17 the compositor needs to manually apply gamma
on `wlr_gamma_control_v1` events. And while we do this, we do not handle
the case correctly where the screen gets turned off and then on again,
in which case the gamma value will not be restored correctly.

Fix this bug by reapplying gamma when powering on the output.