wm/xdg_popup: fix positioning of nested popups
When we create nested popups we add their scene tree to the scene of
their top-level surface. This is wrong as they should be added to the
scene tree of their parent popup, which causes us to position popups at
incorrect positions.

Fix this by adding the popups to their direct parent's scene trees.