wm: implement support for session lock protocol
In the past, locking has been implemented via a combination of the layer
shell and the input inhibit protocols. The latter has been deprecated in
favor of a new session lock protocol.

One of the most important issues that th combination of protocols had is
that they fail open. When the client fails, its surface would be gone
and the input is restored, which essentially unlocks the session again.
The server also cannot second-guess what the client is doing and thus
cannot properly handle this any different

With the new protocol the whole setup is now explicit, and thus it is
possible for the server to stay locked in case the client dies.