wm/xdg_popup: fix handling of destroy events
When creating a popup we register for destroy events of its surface.
This may cause a user-after-free when the surface is destroyed after the
popup and the surface sends another commit event because we'd try to
access the `wlr_xdg_popup` in that case, which may have already been
destroyed. The result is a crash.

Fix this bug by registering for the popup's destroy event instead of
the one from its surface.