wm/xdg_toplevel: fix handling of destroy events
Same as with cb22546 (wm/xdg_popup: fix handling of destroy events,
2024-08-28), we also subscribe to destroy events from the underlying
surface and instead of subscribing to destroys from the XDG popup
itself. As with XDG popups, this can lead to a use-after-free that
ultimately causes us to crash.

Fix this by subscribing to the event from the XDG toplevel, not from its
surface.