wm/popup: fix segfault with layer-shell popups
We register the `new_popup` event exposed by the XDG shell. This event
triggers for both popups with XDG and layer-shell clients as parents.
But in the latter case, those popups don't have the parent pointer set,
which makes it impossible to figure out who has been creating them,
leading to a crash.

Refactor the code so that we instead register the `new_popup` event for
all clients individually. This significantly simplifies the code and
ensures that we have access to the pointer directly.