wm/layer_client: fix focus of newly mapped clients
When a new layer-shell client is being mapped we try to focus it in some
cases. But focussing the client is ineffective because we do so before
recalculating the output, and consequently we haven't yet configured the
layer surface.

Fix this by reordering the code such that adjusting the focus comes
last.