wm/cursor: update focus with active session lock
When the server is locked we don't handle cursor updates, except for
moving the cursor itself. This also includes focus updates, which has
the unfortunate consequence that we may end up with no focussed lock
surface at all, which makes it impossible to unlock the system at all in
the worst case.

Handle focus updates on cursor movements with an active session lock to
improve the situation.