wm: fix crash with multiple idle inhibitors
When there are multiple idle inhibitors we register all of them with the
same listener for destroy events. The consequence is that when we then
unregister the listener we may not remove the expected listener as we
simply unlink the head of that list. This later on triggers an assert,
because wlroots verifies that all listeners for the idle inhibitor got
pruned.

Fix this by allocating a separate structure for every inhibitor that we
can use to register and unregister the event handler.